rpc: identity verification with cached JWKS keys
